test_planner: Log stdout and stderr from repo tool when it fails
Also add a bit more logging. All of this will be useful for debugging
issues in the wild.

The repo invocation is failing locally for me, and this will help me
debug it.

The deploy_cipd update is because I ran ./deploy_cipd.sh locally, so it
updated the latest package in CIPD.

The go.mod update is correct, since that dep is direct.

BUG=chromium:922995
TEST=Local run

Change-Id: I9e3ccb88ab61291812bc42a0c26bf12bad57b1d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/test_planner/+/1565139
Tested-by: Sean Abraham <seanabraham@chromium.org>
Reviewed-by: Evan Hernandez <evanhernandez@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/1679310
Reviewed-by: Sean Abraham <seanabraham@chromium.org>