test_planner: Update test planner to use the new repo and build fields
The serialized build proto field removes the need for writing >100 build
protos to disk (creating a bunch of ugly write_build_ steps in Milo).

The repo tool path is needed because the test planner binary invokes the
repo tool and I'd previously assumed that repo would be on the PATH of
the bots. That wasn't true :/.

BUG=chromium:922995
TEST=unit tests, local run of the tool

Change-Id: I905e7fb5dca9d8451cffb32560f9359346f621d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/test_planner/+/1565926
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Tested-by: Sean Abraham <seanabraham@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/1679312
Reviewed-by: Sean Abraham <seanabraham@chromium.org>