test_planner: Provide useful error message when artifacts not provided
Currently the program panics when the gs_path or gs_bucket aren't
provided: https://screenshot.googleplex.com/K21EmC7Ddsx

I didn't write tests for this, because they'd be more trouble than
they're worth. Even if this doesn't work, all that happens is that
a different error is logged.

(As for why the prod logs reference my username, that's a TODO for me...)

BUG=chromium:922995
TEST=unit tests

Change-Id: I88dc879ecbd3dffedc54fb5b09b07be430714fa6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/test_planner/+/1579478
Reviewed-by: David Burger <dburger@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Tested-by: Sean Abraham <seanabraham@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/1679320
Reviewed-by: Sean Abraham <seanabraham@chromium.org>