pointless_build_checker: Use new build_irrelevance config
This removes the hardcoded paths that are irrelevant to Portage.

This also fixes some questionable logic I was using previously, which
just used prefix-matching on file paths as strings. Now file and
directory paths are handled properly (and "a/b/cccc" isn't regarded as
an irrelevant path just because "a/b/c" is an irrelevant path).

BUG=chromium:956577
TEST=Unit tests, local run

Change-Id: I7ea2beadcf7ce81b201a2531852dd1298269173f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/test_planner/+/1600986
Reviewed-by: Evan Hernandez <evanhernandez@chromium.org>
Tested-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/1679331
Reviewed-by: Sean Abraham <seanabraham@chromium.org>