test_plan_generator: Start writing new TestUnit fields to output
We'll dual-write for now to the new and old places, then migrate recipes
to use the new fields, then drop writing the old test units.

In the mean time, a lot of this code (and tests) look bloated, but it'll
be easy to clean that up afterward.

There's an outstanding issue of figuring out if the builds actually have
the critical field set at all. This CL assumes that the field is set,
and defaults to critical := true anyway, so there's no harm.

BUG=chromium:963016
TEST=unit tests, local run

Change-Id: Ic191f1ad43bbf5218a3bdefe46ba289ae3be3123
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/test_planner/+/1613340
Reviewed-by: Evan Hernandez <evanhernandez@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Tested-by: Sean Abraham <seanabraham@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/1679336
Reviewed-by: Sean Abraham <seanabraham@chromium.org>