testplans: Actually read errors, deflake tests
I was previously forgetting to read the errors from retried calls.
I also previously had a test with two items in a map, and the test would
sometimes fail because of the iteration order of items in the map. The
test only really needs one item in the map.

BUG=None
TEST=unit tests

Change-Id: I77b2a70ee994457940967020e814f5cf5f5bfc22
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/test_planner/+/1613765
Reviewed-by: David Burger <dburger@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Tested-by: Sean Abraham <seanabraham@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/1679339
Reviewed-by: Sean Abraham <seanabraham@chromium.org>