pointless: use input GerritChanges rather than input Build
I'm cutting the pointless build checker's depedence on the input Builds,
since I want to be able to run this program from the orchestrator
itself, in addition to from the child builders.

This also updates a package alias in two cmd packages, since it was
previously unable to compile due to a duplicate "gerrit". I have a fix
coming to run_tests.sh to help prevent such breakages in the future.

BUG=chromium:971464
TEST=run_tests.sh

Change-Id: I31164207581fe0570b5c2234ac2850db450e9f44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/1687920
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Tested-by: Sean Abraham <seanabraham@chromium.org>