Assert functions for test harness
This CL adds various assert functions to the harness that let clients
assert on certain properties of the harness setup.
I'd like to note that repo_harness is intended to be fully generic and
not dependent on the structure of the CrOS project. In subsequent CLs I will
introduce something like cros_repo_harness, which will include cros-specific
assert functions.

TEST=run_tests.sh
BUG=chromium:980346

Change-Id: I0cf2cf63f034ee38df092a2064967c14a25672db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/1709529
Commit-Queue: Jack Neus <jackneus@google.com>
Tested-by: Jack Neus <jackneus@google.com>
Reviewed-by: Evan Hernandez <evanhernandez@chromium.org>