branch_util: refactor version code and fix bug with source branch bumping
Refactored and commented source version branch bumping code to hopefully
be more clear.

When refactored I noticed that when creating a release branch, I always
bump the "master" branch, which is not necessarily the correct revision
for the source branch. Then I noticed I was also erroneously checking
the version from the "master" branch when validating the branch-to-be.
This commit fixes that and adds a functional test to hopefully
prevent such oversights in the future.

BUG=chromium:980346
TEST=run_tests.sh

Change-Id: I8bc6af7527654f63680b3670f8993fae864e3a83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/1764843
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Evan Hernandez <evanhernandez@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: Jack Neus <jackneus@google.com>
Tested-by: Jack Neus <jackneus@google.com>