branch_util: warning instead of error when project dne
We don't really know what to do with a project that doesn't exist in the working manifest,
which is our source of truth. Our best bet is to just use what we have in the manifest
we're repairing.

BUG=None
TEST=run_tests.sh

Change-Id: I8f65097ab87e4d3fe488d365ca9194425dc01dbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/1769384
Tested-by: Jack Neus <jackneus@google.com>
Tested-by: Sean Abraham <seanabraham@chromium.org>
Auto-Submit: Jack Neus <jackneus@google.com>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>