cycler: Add an example single integration test for noop
This commit adds an integration test for the noop effect. This
also will have the chance to test logging, statistics histograms
and other functionality that all runs share.

Its an early version of how we might write the rest of the tests
so I'm putting it out on its own for review and comment without
getting too commited.

BUG=chromium:1035045
TEST=../cmd/cycler/integration_test/test_noop/test_noop.sh

Change-Id: Ia7dfe3b9b2b6bf96276bc080420270ef8198f5fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/1986104
Reviewed-by: David Burger <dburger@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: George Engelbrecht <engeg@google.com>
Tested-by: George Engelbrecht <engeg@google.com>