branch_util: Change how output is generated
This CL adresses the problems outlined in crbug.com/999644.
I changed the way the tool repairs manifests (generates output) so
that it reads in the source manifest without unmarshalling it, makes the
minimal necessary changes, and writes the manifest back to disk. This
preserves element ordering and comments.

BUG=chromium:999644
TEST=tests, still need to local run (but can't because not Googler)

Change-Id: I445d12ac1b9dfc53f7fff8d79d06b0d4987d021b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/2125421
Reviewed-by: Mike Nichols <mikenichols@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Tested-by: Sean Abraham <seanabraham@chromium.org>