cycler: add move_old_paladin configuration
This configuration will be executed to move old paladin
artifacts to the staging bucket. We've verified the
overwhelming number of objects were moved in the last
dup run.

We'll move them to a unique prefix and this will inflate
the backup bucket to 6 PB. Once we're successful and we've
baked a bit, we can delete the original dup run to be
more efficient.

BUG=chromium:1035045
TEST=dup run was successful

Change-Id: I3d114b9a350ed06aa35a2a30a96f182676736953
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/2181537
Commit-Queue: George Engelbrecht <engeg@google.com>
Tested-by: George Engelbrecht <engeg@google.com>
Reviewed-by: David Burger <dburger@chromium.org>