build_plan: only have chromite builders ignore global irrelevance
See the bug for the problem we're currently facing. This not-so-great
solution reflects the fact that our builderconfigs don't tell us at the
moment which builders produce a Chrome OS image (via Portage) and
which are special builders that don't use Portage. The chromite builders
are the only exceptions to this at the moment, so we should just name
them in the code for now. I'm not aware of any other child builders of
the orchestrator coming online that would make this not work.

BUG=chromium:1103256
TEST=unit tests

Change-Id: I002118ba44da098beca4affaa48d4b8ef6384ff5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/2287361
Reviewed-by: Navil Perez <navil@google.com>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Tested-by: Sean Abraham <seanabraham@chromium.org>