branch_util:  Omit branch suffix for branches with multiple manifest entries
I added in functionality to omit the suffix for the main branch
when multiple entries in the manifest exist. This is accomplished
by mapping the main branch with a replacement suffix.

Extra tests were added in the helper_test.go file to check
functionality.

Bug: b/162891513
Test: unit tests; Local dry runs
Change-Id: If72ac832c59706442b94c061e9e084fb9b5ea682
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/go/+/2444011
Reviewed-by: Mike Nichols <mikenichols@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: Julio Hurtado <juahurta@google.com>
Tested-by: Julio Hurtado <juahurta@google.com>